Pull sed parser into new file
