+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/treeSitterCommandParser.ts
@@ -9,6 +9,8 @@ import { BugIndicatingError, ErrorNoTelemetry } from '../../../../../base/common
 import { Lazy } from '../../../../../base/common/lazy.js';
 import { Disposable, MutableDisposable, toDisposable } from '../../../../../base/common/lifecycle.js';
 import { ITreeSitterLibraryService } from '../../../../../editor/common/services/treeSitter/treeSitterLibraryService.js';
+import { ICommandFileWriteParser } from './commandParsers/commandFileWriteParser.js';
+import { SedFileWriteParser } from './commandParsers/sedFileWriteParser.js';
 
 export const enum TreeSitterCommandParserLanguage {
 	Bash = 'bash',
@@ -18,6 +20,9 @@ export const enum TreeSitterCommandParserLanguage {
 export class TreeSitterCommandParser extends Disposable {
 	private readonly _parser: Lazy<Promise<Parser>>;
 	private readonly _treeCache = this._register(new TreeCache());
+	private readonly _commandFileWriteParsers: ICommandFileWriteParser[] = [
+		new SedFileWriteParser(),
+	];
 
 	constructor(
 		@ITreeSitterLibraryService private readonly _treeSitterLibraryService: ITreeSitterLibraryService,
@@ -62,11 +67,12 @@ export class TreeSitterCommandParser extends Disposable {
 	}
 
 	/**
-	 * Extracts file targets from `sed` commands that use in-place editing (`-i`, `-I`, or `--in-place`).
+	 * Extracts file targets from commands that perform file writes beyond shell redirections.
+	 * Uses registered command parsers (e.g., for `sed -i`) to detect command-specific file writes.
 	 * Returns an array of file paths that would be modified.
 	 */
-	async getSedInPlaceFiles(languageId: TreeSitterCommandParserLanguage, commandLine: string): Promise<string[]> {
-		// This is only relevant for bash-like shells
+	async getCommandFileWrites(languageId: TreeSitterCommandParserLanguage, commandLine: string): Promise<string[]> {
+		// Currently only bash-like shells are supported for command-specific parsing
 		if (languageId !== TreeSitterCommandParserLanguage.Bash) {
 			return [];
 		}
@@ -78,203 +84,13 @@ export class TreeSitterCommandParser extends Disposable {
 		const result: string[] = [];
 		for (const capture of captures) {
 			const commandText = capture.node.text;
-			const sedFiles = this._parseSedInPlaceFiles(commandText);
-			result.push(...sedFiles);
+			for (const parser of this._commandFileWriteParsers) {
+				if (parser.canHandle(commandText)) {
+					result.push(...parser.extractFileWrites(commandText));
+				}
+			}
 		}
 		return result;
-	}
-
-	/**
-	 * Parses a sed command to extract files being edited in-place.
-	 * Handles:
-	 * - `sed -i 's/foo/bar/' file.txt` (GNU)
-	 * - `sed -i.bak 's/foo/bar/' file.txt` (GNU with backup suffix)
-	 * - `sed -i '' 's/foo/bar/' file.txt` (macOS/BSD with empty backup suffix)
-	 * - `sed --in-place 's/foo/bar/' file.txt` (GNU long form)
-	 * - `sed --in-place=.bak 's/foo/bar/' file.txt` (GNU long form with backup)
-	 * - `sed -I 's/foo/bar/' file.txt` (BSD case-insensitive variant)
-	 */
-	private _parseSedInPlaceFiles(commandText: string): string[] {
-		// Check if this is a sed command with in-place flag
-		const sedMatch = commandText.match(/^sed\s+/);
-		if (!sedMatch) {
-			return [];
-		}
-
-		// Check for -i, -I, or --in-place flag
-		const inPlaceRegex = /(?:^|\s)(-[a-zA-Z]*[iI][a-zA-Z]*\S*|--in-place(?:=\S*)?|(-i|-I)\s*'[^']*'|(-i|-I)\s*"[^"]*")(?:\s|$)/;
-		if (!inPlaceRegex.test(commandText)) {
-			return [];
-		}
-
-		// Parse the command to extract file arguments
-		// We need to skip: the 'sed' command, flags, and sed scripts/expressions
-		const tokens = this._tokenizeSedCommand(commandText);
-		return this._extractSedFileTargets(tokens);
-	}
-
-	/**
-	 * Tokenizes a sed command into individual arguments, handling quotes and escapes.
-	 */
-	private _tokenizeSedCommand(commandText: string): string[] {
-		const tokens: string[] = [];
-		let current = '';
-		let inSingleQuote = false;
-		let inDoubleQuote = false;
-		let escaped = false;
-
-		for (let i = 0; i < commandText.length; i++) {
-			const char = commandText[i];
-
-			if (escaped) {
-				current += char;
-				escaped = false;
-				continue;
-			}
-
-			if (char === '\\' && !inSingleQuote) {
-				escaped = true;
-				current += char;
-				continue;
-			}
-
-			if (char === '\'' && !inDoubleQuote) {
-				inSingleQuote = !inSingleQuote;
-				current += char;
-				continue;
-			}
-
-			if (char === '"' && !inSingleQuote) {
-				inDoubleQuote = !inDoubleQuote;
-				current += char;
-				continue;
-			}
-
-			if (/\s/.test(char) && !inSingleQuote && !inDoubleQuote) {
-				if (current) {
-					tokens.push(current);
-					current = '';
-				}
-				continue;
-			}
-
-			current += char;
-		}
-
-		if (current) {
-			tokens.push(current);
-		}
-
-		return tokens;
-	}
-
-	/**
-	 * Extracts file targets from tokenized sed command arguments.
-	 * Files are generally the last non-option, non-script arguments.
-	 */
-	private _extractSedFileTargets(tokens: string[]): string[] {
-		if (tokens.length === 0 || tokens[0] !== 'sed') {
-			return [];
-		}
-
-		const files: string[] = [];
-		let i = 1; // Skip 'sed'
-		let foundScript = false;
-
-		while (i < tokens.length) {
-			const token = tokens[i];
-
-			// Long options
-			if (token.startsWith('--')) {
-				if (token === '--in-place' || token.startsWith('--in-place=')) {
-					// In-place flag (already verified we have one)
-					i++;
-					continue;
-				}
-				if (token === '--expression' || token === '--file') {
-					// Skip the option and its argument
-					i += 2;
-					foundScript = true;
-					continue;
-				}
-				if (token.startsWith('--expression=') || token.startsWith('--file=')) {
-					i++;
-					foundScript = true;
-					continue;
-				}
-				// Other long options like --sandbox, --debug, etc.
-				i++;
-				continue;
-			}
-
-			// Short options
-			if (token.startsWith('-') && token.length > 1 && token[1] !== '-') {
-				// Could be combined flags like -ni or -i.bak
-				const flags = token.slice(1);
-
-				// Check if this is -i with backup suffix attached (e.g., -i.bak)
-				const iIndex = flags.indexOf('i');
-				const IIndex = flags.indexOf('I');
-				const inPlaceIndex = iIndex >= 0 ? iIndex : IIndex;
-
-				if (inPlaceIndex >= 0 && inPlaceIndex < flags.length - 1) {
-					// -i.bak style - backup suffix is attached
-					i++;
-					continue;
-				}
-
-				// Check if -i or -I is the last flag and next token could be backup suffix
-				if ((flags.endsWith('i') || flags.endsWith('I')) && i + 1 < tokens.length) {
-					const nextToken = tokens[i + 1];
-					// macOS/BSD style: -i '' or -i "" (empty string backup suffix)
-					if (nextToken === '\'\'' || nextToken === '""' || (nextToken.startsWith('\'') && nextToken.endsWith('\'')) || (nextToken.startsWith('"') && nextToken.endsWith('"'))) {
-						i += 2;
-						continue;
-					}
-				}
-
-				// Check for -e or -f which take arguments
-				if (flags.includes('e') || flags.includes('f')) {
-					const eIndex = flags.indexOf('e');
-					const fIndex = flags.indexOf('f');
-					const optIndex = eIndex >= 0 ? eIndex : fIndex;
-
-					// If -e or -f is not the last character, the rest of the token is the argument
-					if (optIndex < flags.length - 1) {
-						foundScript = true;
-						i++;
-						continue;
-					}
-
-					// Otherwise, the next token is the argument
-					foundScript = true;
-					i += 2;
-					continue;
-				}
-
-				i++;
-				continue;
-			}
-
-			// Non-option argument
-			if (!foundScript) {
-				// First non-option is the script (unless -e/-f was used)
-				foundScript = true;
-				i++;
-				continue;
-			}
-
-			// Subsequent non-option arguments are files
-			// Strip surrounding quotes from file path
-			let file = token;
-			if ((file.startsWith('\'') && file.endsWith('\'')) || (file.startsWith('"') && file.endsWith('"'))) {
-				file = file.slice(1, -1);
-			}
-			files.push(file);
-			i++;
-		}
-
-		return files;
 	}
 
 	private async _queryTree(languageId: TreeSitterCommandParserLanguage, commandLine: string, querySource: string): Promise<QueryCapture[]> {
